Extract user avatar to a separate component
